\name{ExploreDatabase}

\alias{ExploreDatabase}

\title{Explore Database Connection}

\description{
A \acronym{GUI} for exploring the database connection
}

\usage{
ExploreDatabase(con, parent = NULL)
}

\arguments{
\item{con}{RODBC; connection object.}
\item{parent}{tkwin; the parent window (optional).}
}

\details{
Something.
}

\author{Fisher, J.C.}

\seealso{
\code{odbcConnect}, \code{sqlTables}
}

\examples{
print("something")
}

\keyword{misc}
