% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRNWIS.R
\name{OpenRNWIS}
\alias{OpenRNWIS}
\title{Open Main Graphical User Interface}
\usage{
OpenRNWIS()
}
\description{
This function activates the main graphcial user interface (GUI) for the \pkg{RNWIS} package.
}
\details{
All functions within the \pkg{RNWIS} package may be called from this GUI.
}
\examples{
\dontrun{
  OpenRNWIS()
}

}
\seealso{
\code{\link{ExploreDatabase}}, \code{\link{MapSites}}, \code{\link{QueryDatabase}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
