% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreDatabase.R
\name{ExploreDatabase}
\alias{ExploreDatabase}
\title{Explore Database Connection}
\usage{
ExploreDatabase(channel, parent = NULL)
}
\arguments{
\item{channel}{RODBC.
Connection to the ODBC database.}

\item{parent}{tkwin.
Parent window (optional)}
}
\description{
A graphical user interface for exploring the contents of a database.
}
\details{
Shows database table names and the column structure for each table.
}
\examples{
\dontrun{
  channel <- RODBC::odbcConnect("NWIS Idaho", uid = "", pwd = "")
  ExploreDatabase(channel)
  close(channel)
}

}
\seealso{
\code{\link{sqlTables}}, \code{\link{sqlColumns}}, \code{\link{sqlPrimaryKeys}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
