\name{MapSites}

\alias{MapSites}

\title{Map Site Locations in Google Maps}

\description{
Something
}

\usage{
MapSites(sites, polygons = NULL, map.id = NULL)
}

\arguments{
\item{sites}{data.frame; something.}
\item{polygons}{list; something.}
\item{map.id}{character; something.}
}

\details{
Something.
}

\author{Fisher, J.C.}

\seealso{
\code{\link{browseURL}}, \code{\link{tempfile}}
}

\examples{
print("something")
}

\keyword{aplot}
