% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapSites.R
\name{MapSites}
\alias{MapSites}
\title{Map Site Locations in Google Maps}
\usage{
MapSites(sites, polygons = NULL, map.id = "map")
}
\arguments{
\item{sites}{data.frame.
Site information, see 'Details'.}

\item{polygons}{gpc.poly.
Polygon information}

\item{map.id}{character.
Unique identifier in map URL}
}
\value{
Writes a \href{https://www.json.org/}{JSON} data file (\file{.json}) to a temporary directory.
  An Rhttpd object is created and started in the internal web server.
}
\description{
Add site markers and polygon objects to \href{https://www.google.com/maps}{Google Maps}.
}
\details{
The \code{sites} data table has components of \code{lat}, \code{lng},
  \code{alt}, \code{site}, \code{name}, \code{agency}, and \code{type}.
  Where \code{lat} and \code{lng} are the latitude and longitude based on the
  \href{https://en.wikipedia.org/wiki/World_Geodetic_System#WGS84}{WGS84} datum;
  \code{alt} is the altitude of the site referenced to the specified vertical datum
  (\href{https://en.wikipedia.org/wiki/Sea_Level_Datum_of_1929}{NGVD29} or \href{https://en.wikipedia.org/wiki/North_American_Vertical_Datum_of_1988}{NAVD 88});
  \code{site} is the unique site identification number;
  \code{name} is the name of the site;
  \code{agency} is the code for the agency reporting the data; and
  \code{type} is the hydrologic setting of the site.
}
\examples{
\dontrun{
  s <- data.frame(lat    = c(43.5757402325, 43.5803046219, 43.8642317971),
                  lng    = c(-112.934719019, -112.876178537, -112.742880208),
                  site   = c(433433112560201, 433449112523101, 435152112443101),
                  name   = c("CPP 1", "NPR Test", "ANP 6"),
                  agency = "USGS", type = "GW", stringsAsFactors = FALSE)
  MapSites(sites = s, map.id = "map01")

  f <- system.file("extdata/ExamplePolygon.ply", package = "RNWIS")
  p <- rgeos::read.polyfile(f, nohole = FALSE)
  MapSites(sites = s, polygons = p, map.id = "map02")
}

}
\seealso{
\code{\link{Rhttpd-class}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{aplot}
