\name{MapSites}

\alias{MapSites}

\title{Map Site Locations in Google Maps}

\description{
Something
}

\usage{
MapSites(data, lat.var, lng.var, site.var, name.var, agency.var, 
         map.id = NULL)
}

\arguments{
\item{data}{data.frame; something.}
\item{lat.var}{character; something.}
\item{lng.var}{character; something.}
\item{site.var}{character; something.}
\item{name.var}{character; something.}
\item{agency.var}{character; something.}
\item{map.id}{character; something.}
}

\details{
Something.
}

\author{Fisher, J.C.}

\seealso{
\code{\link{browseURL}}, \code{\link{tempfile}}
}

\examples{
print("something")
}

\keyword{aplot}
