\name{OpenRNWIS}

\alias{OpenRNWIS}

\title{Open Main Graphical User Interface}

\description{
This function activates the main \acronym{GUI} for \pkg{RNWIS}.
}

\usage{
OpenRNWIS()
}

\details{
All functions within \pkg{RNWIS} are accessible in this \acronym{GUI}.
}

\author{Fisher, J.C.}

\examples{
\donttest{OpenRNWIS()}
}

\keyword{misc}
