\name{OpenRNWIS}

\alias{OpenRNWIS}

\title{Open Main Graphical User Interface}

\description{
This function activates the main \acronym{GUI} for \pkg{RNWIS}.
}

\usage{
OpenRNWIS()
}

\details{
All functions within \pkg{RNWIS} are accessible in this \acronym{GUI}.
}

\author{Fisher, J.C.}

\seealso{
\code{odbcConnect}, \code{writeOGR}, \code{\link{write.table}},
\code{\link{point.in.polygon}}
}

\examples{
\donttest{OpenRNWIS()}
}

\keyword{misc}
