% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatabase.R
\name{QueryDatabase}
\alias{QueryDatabase}
\title{Query the National Water Information System Database}
\usage{
QueryDatabase(
  channel,
  sqtable,
  sqvars = "*",
  site.no.var = NULL,
  site.no = NULL,
  site.tp.cd.var = NULL,
  site.tp.cd = NULL,
  agency.cd.var = NULL,
  agency.cd = NULL,
  lat.var = NULL,
  lat.lim = c(NA, NA),
  lng.var = NULL,
  lng.lim = c(NA, NA),
  alt.var = NULL,
  alt.lim = c(NA, NA),
  d.t.var = NULL,
  d.t.lim = c(NA, NA)
)
}
\arguments{
\item{channel}{RODBC.
A connection to a ODBC database.}

\item{sqtable}{character.
Name of the table from which data is to be retrieved.}

\item{sqvars}{character.
Vector of column names in queried table to be included in the final query results;
its default is an asterisk \file{*} and specifies that all columns will be returned.}

\item{site.no.var}{character.
Column name of table which shows the site identification number.}

\item{site.no}{numeric.
Vector of site identification numbers}

\item{site.tp.cd.var}{character.
Column name of queried table which shows the site type.}

\item{site.tp.cd}{character.
Vector of site type codes}

\item{agency.cd.var}{character.
Column name of queried table which shows the agency code.}

\item{agency.cd}{character.
Vector of agency codes}

\item{lat.var}{character.
Column name of queried table which shows the latitude.}

\item{lat.lim}{numeric.
Vector of minimum and maximum latitude values (WGS84).}

\item{lng.var}{character.
Column name of queried table which shows the longitude.}

\item{lng.lim}{numeric.
Vector of minimum and maximum longitude values (WGS84).}

\item{alt.var}{character.
Column name of queried table which shows the altitude.}

\item{alt.lim}{numeric.
Vector of minimum and maximum altitude values (NGVD29 or NAVD 88).}

\item{d.t.var}{character.
Column name of queried table which shows a date and time variable.}

\item{d.t.lim}{POSIXt.
Vector of minimum and maximum date values}
}
\value{
On success, returns an object of class \code{data.frame} or \code{character}.
  On failure, returns a vector of error message(s).
}
\description{
This function acquire data from a single database table using a
site selection criteria to constrain the number of sites selected.
}
\examples{
\dontrun{
  con <- RODBC::odbcConnect("NWIS Idaho", uid = "", pwd = "")

  d <- QueryDatabase(con, sqtable = "sitefile_01",
                     sqvars = c("site_no", "alt_va"),
                     site.no.var = "site_no",
                     site.no = c(432700112470801, 435038112453401))

  d <- QueryDatabase(con, sqtable = "sitefile_01",
                     sqvars = c("site_no", "dec_lat_va", "dec_long_va"),
                     site.tp.cd.var = "site_tp_cd", site.tp.cd = "GW",
                     agency.cd.var = "agency_cd", agency.cd = c("USGS", "USEPA"),
                     lng.var = "dec_long_va", lng.lim = c(-114.00, -112.00),
                     lat.var = "dec_lat_va", lat.lim = c(43.00, 44.00))

  d <- QueryDatabase(con, sqtable = "gw_lev_01",
                     sqvars = c("site_no", "lev_dt", "lev_va"),
                     site.no.var = "site_no", site.no = 432700112470801)

  close(con)
}

}
\seealso{
\code{\link{sqlQuery}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{IO}
